Light refactor to initToken dir.
